+++ sdmx-dl-provider-ri/src/test/java/internal/sdmxdl/provider/ri/web/monitors/UpptimeIdTest.java
@@ -27,6 +27,15 @@ public class UpptimeIdTest {
     }
 
     @Test
+    public void toReportURL() throws MalformedURLException {
+        assertThat(UpptimeId.builder().owner("").repo("").site("").build().toReportURL())
+                .isEqualTo(new URL("https://.github.io//history/"));
+
+        assertThat(UpptimeId.builder().owner("nbbrd").repo("sdmx-upptime").site("ECB").build().toReportURL())
+                .isEqualTo(new URL("https://nbbrd.github.io/sdmx-upptime/history/ecb"));
+    }
+
+    @Test
     public void parse() {
         assertThat(UpptimeId.parse(URI.create("upptime:/nbbrd/sdmx-upptime/ECB")))
                 .isEqualTo(UpptimeId.builder().owner("nbbrd").repo("sdmx-upptime").site("ECB").build());